FIXED: Codestyle janitored some warnings, this early morning (14:34). It seems I need some more Eyes Cold Lemonade: http://www.youtube.com/watch?v=165VjNKRNdw

+++ src/arcemu-world/MailSystem.cpp
@@ -881,7 +881,7 @@ void Mailbox::Load(QueryResult * result)
 	uint32 i;
 	char * str;
 	char * p;
-	uint64 itemguid;
+	uint32 itemguid;
 
 	do 
 	{
